requirements.txt
PyYAML==6.0.2
typing_extensions==4.14.1
xmltodict==0.14.2
